Split MAD to new file
